pyproject.toml
[tool.poetry]
name = "borderline"
version = "0.1.0"
description = "Borderline app"
authors = ["Your Name <you@example.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13"
opencv-python = "^4.10.0.84"
imutils = "^0.5.4"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
